requirements.txt
django-quill-editor==0.1.40